==> E - SBERT Model Training/14_GC_EmbeddingJobAds.ipynb <==
# Imports

# Mount Google Drive (if using Google Colab)
# from google.colab import drive
# drive.mount('/content/drive')
# %cd /content/drive/MyDrive/TRAINING

# Install necessary libraries (if using Google Colab)
# %%capture
# !pip install sentence_transformers

import sys
sys.path.append("..")
from _utils import load_json,load_pickle, write_pickle
from training_helpers import encode_jobs

# Import necessary libraries
from sentence_transformers import SentenceTransformer, InputExample, evaluation, models, losses, util
from torch.utils.data import DataLoader
from tqdm import tqdm
import random
import numpy as np
import json
import pandas as pd
import nltk
nltk.download('punkt')
from collections import Counter
import os
from datetime import datetime
import torch

# Helpers

# Community detection function for clustering embeddings
def community_detection(embeddings, threshold=0.75, min_community_size=10, batch_size=1024):
    """
    Perform fast community detection on embeddings.
    Args:
        embeddings (Tensor): Input embeddings.
        threshold (float): Cosine similarity threshold for clustering.
        min_community_size (int): Minimum size of a community.
        batch_size (int): Batch size for processing.
    Returns:
        list: List of unique communities.
    """
    threshold = torch.tensor(threshold, device=embeddings.device)
    extracted_communities = []
    min_community_size = min(min_community_size, len(embeddings))
    sort_max_size = min(max(2 * min_community_size, 50), len(embeddings))
    for start_idx in range(0, len(embeddings), batch_size):
        cos_scores = util.cos_sim(embeddings[start_idx:start_idx + batch_size], embeddings)
        top_k_values, _ = cos_scores.topk(k=min_community_size, largest=True)
        for i in range(len(top_k_values)):
            if top_k_values[i][-1] >= threshold:
                new_cluster = []
                top_val_large, top_idx_large = cos_scores[i].topk(k=sort_max_size, largest=True)
                while top_val_large[-1] > threshold and sort_max_size < len(embeddings):
                    sort_max_size = min(2 * sort_max_size, len(embeddings))
                    top_val_large, top_idx_large = cos_scores[i].topk(k=sort_max_size, largest=True)
                for idx, val in zip(top_idx_large.tolist(), top_val_large):
                    if val < threshold:
                        break
                    new_cluster.append(idx)
                extracted_communities.append(new_cluster)
        del cos_scores
    extracted_communities = sorted(extracted_communities, key=lambda x: len(x), reverse=True)
    unique_communities = []
    extracted_ids = set()
    for cluster_id, community in enumerate(extracted_communities):
        community = sorted(community)
        non_overlapped_community = []
        for idx in community:
            if idx not in extracted_ids:
                non_overlapped_community.append(idx)
        if len(non_overlapped_community) >= min_community_size:
            unique_communities.append(non_overlapped_community)
            extracted_ids.update(non_overlapped_community)
    unique_communities = sorted(unique_communities, key=lambda x: len(x), reverse=True)
    return unique_communities

# Filter communities based on ESCO job IDs
def filter_communities_total(df, columnname_to_encode, model, min_community_size=10, threshold=0.75):
    """
    Filter communities for each ESCO job ID.
    Args:
        df (DataFrame): Input DataFrame.
        columnname_to_encode (str): Column name to encode.
        model (SentenceTransformer): Pre-trained model.
        min_community_size (int): Minimum community size.
        threshold (float): Cosine similarity threshold.
    Returns:
        DataFrame: DataFrame with community assignments.
    """
    unique_texts = list(set(df[columnname_to_encode]))
    print(f"{len(unique_texts)} ads to encode.")
    embedding_map = {}
    embeddings = model.encode(unique_texts, show_progress_bar=True, convert_to_tensor=True, batch_size=64)
    print("Finished encoding.")
    for text, embedding in zip(unique_texts, embeddings.tolist()):
        embedding_map[text] = embedding
    df["embeddings"] = df[columnname_to_encode].map(embedding_map)
    results = []
    esco_ids = list(df["esco_id"].unique())
    df["community"] = None
    print("Adding Communities.")
    for id in tqdm(esco_ids):
        filtered_df = df[df["esco_id"] == id].reset_index(drop=True)
        if len(filtered_df) < min_community_size:
            filtered_df["community"] = 0
        else:
            embds = [torch.FloatTensor(x) for x in filtered_df["embeddings"]]
            embds = torch.stack(embds)
            communities = community_detection(embds, threshold=threshold, min_community_size=min_community_size)
            if len(communities) > 0:
                for index, community in enumerate(communities):
                    filtered_df.loc[community, "community"] = index
            else:
                filtered_df["community"] = 0
        results += filtered_df.to_dict("records")
    result_df = pd.DataFrame(results)
    result_df.reset_index(drop=True, inplace=True)
    return result_df

# Model Loading

# Load Data

# Load job advertisements data
ads_df = pd.DataFrame(load_json(r"../00_data/EURES/0_pars_short_ads_final.json"))
ads_df.columns

# Concatenate title and short texts for encoding
def concat_short(ad):
    return ad["title"] + ad["short_texts"]

# Apply concatenation to create a new column
ads_df["short+title"] = ads_df.apply(concat_short, axis=1)

# Count unique ESCO job IDs
len(ads_df["esco_id"].unique())

# Load ESCO job data and create a lookup dictionary
jobs = load_json("../00_data/ESCO/ESCO_JOBS_ALL.json")
esco_lookup = {job["jobid_esco"]: job["jobtitle"] for job in jobs}

# Encoding

# Define model paths for encoding
paths = [
    "../00_data/SBERT_Models/models/gbert_batch32_woTSDAE_2e-05_f10/",
    "../00_data/SBERT_Models/models/jobgbert_batch32_woTSDAE_2e-05_f10/",
    "../00_data/SBERT_Models/models/consultantbert_multilingual_best/"
]

# Encode job advertisements and save embeddings
for model_path in paths:
    model = SentenceTransformer(model_path)
    embeddings = encode_jobs(model)
    write_pickle(f"{model_path}embeddings.pkl", embeddings)

# Create centroids for job advertisements and save them
for model_path in paths:
    print(f"Creating Encodings for {model_path}")
    model = SentenceTransformer(model_path)
    df_new = filter_communities_total(ads_df, "short+title", model, min_community_size=100, threshold=0.5)
    unique_ids = set(df_new["esco_id"].unique())
    unfiltered_centroids_dict = {}
    filtered_centroids_dict = {}
    print("Creating Centroids.")
    for id in tqdm(unique_ids):
        id_filter = df_new[df_new["esco_id"] == id]
        unfiltered_centroids = np.stack(list(id_filter["embeddings"])).mean(axis=0, dtype="float32")
        unfiltered_centroids_dict[id] = unfiltered_centroids
        id_filter = df_new[(df_new["esco_id"] == id) & (df_new["community"] == 0)]
        filtered_centroids = np.stack(list(id_filter["embeddings"])).mean(axis=0, dtype="float32")
        filtered_centroids_dict[id] = filtered_centroids
    write_pickle(f"{model_path}centroids_unfiltered.pkl", unfiltered_centroids_dict)
    write_pickle(f"{model_path}centroids_filtered.pkl", filtered_centroids_dict)

# Transform embeddings and save mappings
for model_path in paths:
    embeddings = load_pickle(f"{model_path}embeddings.pkl")
    print(embeddings.keys())
    for centroid_path in ["centroids_filtered.pkl", "centroids_unfiltered.pkl"]:
        centroids = load_pickle(f"{model_path}{centroid_path}")
        centroids_ids = list(centroids.keys())
        centroid_values = list(centroids.values())
        centroid_jobs = [esco_lookup[x] for x in centroids_ids]
        centroid_kind = "adcentroid_" + centroid_path.split("_")[1].split(".")[0]
        embeddings[centroid_kind] = {"jobtitle": centroid_jobs, "esco_id": centroids_ids, "embeddings": centroid_values}
    job_centroid = []
    for k in ['desc', 'adcentroid_unfiltered']:
        for id, job, embedding in zip(embeddings[k]["esco_id"], embeddings[k]["jobtitle"], embeddings[k]["embeddings"]):
            job_centroid.append({"esco_id": id, "jobtitle": job, "embeddings": embedding})
    job_embeddings_jobtitle, job_embeddings_esco_ids, job_embeddings = [], [], []
    centroid_df = pd.DataFrame(job_centroid)
    for id in centroid_df["esco_id"].unique():
        filtered_df = centroid_df[centroid_df["esco_id"] == id]
        stacked_embedding = np.stack(list(filtered_df["embeddings"])).mean(axis=0, dtype="float32")
        job_embeddings.append(stacked_embedding)
        job_embeddings_esco_ids.append(id)
        job_embeddings_jobtitle.append(filtered_df["jobtitle"].iloc[0])
    embeddings["job_centroid"] = {"jobtitle": job_embeddings_jobtitle, "esco_id": job_embeddings_esco_ids, "embeddings": job_embeddings}
    write_pickle(f"{model_path}embeddings.pkl", embeddings)
